rearranged classes to be more readable
